test(feature): hoist small_gdf/lfc fixtures to feature/lazy/conftest.py (Theme D.2)
Move the byte-identical small_gdf/lfc fixtures (duplicated in test_collection.py and
test_to_parquet.py) into a new conftest; drop the now-unused geopandas/shapely imports.
The LazyFeatureCollection import stays inline under the importorskip guard. No logic
changes.
