+++ tests/feature/lazy/test_collection.py
@@ -4,33 +4,13 @@ from __future__ import annotations
 
 import pickle
 
-import geopandas as gpd
 import pytest
-from shapely.geometry import Point
 
 from pyramids.base.protocols import LazySpatialObject, SpatialObject, is_lazy
 from pyramids.feature import FeatureCollection
 from tests._marks import requires_dask_geopandas
 
 pytestmark = pytest.mark.parquet_lazy
-
-
-@pytest.fixture
-def small_gdf():
-    return gpd.GeoDataFrame(
-        {"id": list(range(10)), "class": ["a"] * 5 + ["b"] * 5},
-        geometry=[Point(i, i) for i in range(10)],
-        crs="EPSG:4326",
-    )
-
-
-@pytest.fixture
-def lfc(small_gdf):
-    dg = pytest.importorskip("dask_geopandas")
-    from pyramids.feature import LazyFeatureCollection
-
-    ddf = dg.from_geopandas(small_gdf, npartitions=2)
-    return LazyFeatureCollection.from_dask_gdf(ddf)
 
 
 @requires_dask_geopandas
